Keep one of the duplicated secrets
There is really no problem with duplicated secrets: these exist, and will be shared
by the ingresses that refer to them. The existing code excluded a secret as soon as
it was referenced by more one than place, but we should really just de-duplicate
the given list based on namespace/name.